app/requirements.txt
fastapi~=0.111.1
uvicorn~=0.30.2
tensorflow~=2.17.0
numpy~=1.26.4
pydantic~=2.8.2
keras~=3.4.1
pillow==10.4.0